+++ castle/cms/browser/content/__init__.py
@@ -3,6 +3,7 @@ import logging
 import math
 import os
 import plone.api as api
+import re
 import shutil
 import tempfile
 import time
@@ -14,6 +15,7 @@ from Acquisition import aq_parent
 from castle.cms import cache
 from castle.cms import commands
 from castle.cms import utils
+from castle.cms.archival import SubrequestUrlOpener
 from castle.cms.browser.utils import Utils
 from castle.cms.commands import exiftool
 from castle.cms.commands import qpdf
@@ -740,26 +742,123 @@ class WorkflowPermissionChecker(object):
         return True
 
 
+class BackendUrlUtils(object):
+    # start of string
+    # optional http:// or https:// is catured as 'protocol'
+    # optional leading slashes are captured as 'leading_slashes'
+    # everything else is captured as 'url'
+    URL_PATTERN = r'^(?P<protocol>https?:\/\/)?(?P<leading_slashes>\/+)?(?P<url>.*)$'
+    UNSET = object()
+    _public_url = UNSET
+    _backend_urls = UNSET
+    _invalid_backend_urls = UNSET
+
+    @property
+    def backend_urls(self):
+        if getattr(self, '_backend_urls', self.UNSET) is self.UNSET:
+            formatted_backend_urls = []
+            backend_urls = api.portal.get_registry_record('plone.backend_url', default=[]) or []
+            for backend_url in backend_urls:
+                formatted_url = self.get_formatted_url(backend_url)
+                if formatted_url and isinstance(formatted_url, basestring):
+                    formatted_backend_urls.append(formatted_url)
+            self._backend_urls = formatted_backend_urls
+        return self._backend_urls
+
+    @property
+    def public_url(self):
+        if getattr(self, '_public_url', self.UNSET) is self.UNSET:
+            public_url = api.portal.get_registry_record('plone.public_url', default=None) or None
+            if public_url and isinstance(public_url, basestring):
+                self._public_url = public_url
+            else:
+                self._public_url = None
+        return self._public_url
+
+    @property
+    def invalid_backend_urls(self):
+        # if a site has the same backend and public url, do not consider it invalid
+        if getattr(self, '_invalid_backend_urls', self.UNSET) is self.UNSET:
+            public_url = self.public_url
+            backend_urls = self.backend_urls
+            if public_url is None:
+                self._invalid_backend_urls = backend_urls
+            else:
+                self._invalid_backend_urls = [
+                    backend_url
+                    for backend_url in backend_urls
+                    if backend_url != public_url
+                ]
+        return self._invalid_backend_urls
+
+    def get_formatted_url(self, url):
+        try:
+            # probably the only way to not get a match here is a non-string
+            match = re.match(self.URL_PATTERN, url)
+            return match.groupdict()['url']
+        except Exception:
+            return None
+
+    def get_invalid_backend_urls_found(self, content):
+        return [
+            invalid_backend_url
+            for invalid_backend_url in self.invalid_backend_urls
+            if invalid_backend_url in content
+        ]
+
+
 class QualityCheckContent(BrowserView):
+    QUALITY_CHECK_URL = '/@@quality-check'
 
-    def __call__(self):
+    @property
+    def formatted_url(self):
+        try:
+            end_index = None
+            if self.request.URL.endswith(self.QUALITY_CHECK_URL):
+                end_index = -1 * len(self.QUALITY_CHECK_URL)
+            return self.request.URL[:end_index]
+        except Exception:
+            return None
 
-        valid = True
+    @property
+    def subrequest_results(self):
+        opener = SubrequestUrlOpener(
+            site=api.portal.get(),
+            check_blacklist=False,
+        )
+        return opener(self.formatted_url, require_public_url=False)
+
+    @property
+    def contains_backend_urls(self):
+        subrequest_results = self.subrequest_results
+        subrequest_html = subrequest_results['data']
+        backend_utils = BackendUrlUtils()
+        backend_urls_found = backend_utils.get_invalid_backend_urls_found(subrequest_html)
+        if len(backend_urls_found) > 0:
+            logger.warn('There were backend urls found in the html')
+            logger.info('Backend urls found: {}'.format(repr(backend_urls_found)))
+            logger.info('Data searched for: ' + self.formatted_url)
+            return True
+        return False
+
+    @property
+    def are_links_valid(self):
         for link in getOutgoingLinks(self.context):
             state = api.content.get_state(obj=link.to_object, default='published')
             if state != 'published':
-                valid = False
-                break
+                return False
+        return True
 
-        headers_ordered = True
-
+    @property
+    def html(self):
         try:
             feed = SearchFeed(api.portal.get())
             adapter = queryMultiAdapter((self.context, feed), IFeedItem)
-            html = adapter.render_content_core().strip()
+            return adapter.render_content_core().strip()
         except Exception:
-            html = ''
+            return ''
 
+    def are_headers_ordered(self, html):
         if html:
             dom = fromstring(html)
             last = 1
@@ -768,19 +867,22 @@ class QualityCheckContent(BrowserView):
                 if idx - last > 1:
                     # means they skipped from say h1 -> h5
                     # h1 -> h2 is allowed
-                    headers_ordered = False
-                    break
+                    return False
                 last = idx
+        return True
 
+    def __call__(self):
+        html = self.html
         self.request.response.setHeader('Content-type', 'application/json')
         return json.dumps({
             'title': self.context.Title(),
             'id': self.context.getId(),
             'description': self.context.Description(),
-            'linksValid': valid,
-            'headersOrdered': headers_ordered,
+            'linksValid': self.are_links_valid,
+            'headersOrdered': self.are_headers_ordered(html),
             'html': html_parser.unescape(html),
             'isTemplate': self.context in get_template_repository_info()['templates'],
+            'containsBackendUrls': self.contains_backend_urls,
         })